pyproject.toml
[project]
name = "publicdata-ca"
version = "0.1.0"
description = "Canadian public data acquisition and processing package"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "pyyaml>=6.0",
    "scikit-learn>=1.3.0",
    "pyarrow>=14.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "pytest>=7.4.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.pytest.ini_options]
testpaths = ["case_studies/hospital_anomalies/tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
